remove freezed & sanitize disciplines/overview
